refactor(logging): avoid exposing flush to downstream consumers
